Update overview-of-the-upcoming-changes-to-office-365-proplus-update-management.md
The changes I am proposing address the fact that much of the working is out of date since it appears the last revision was in October of 2017.  The changes attempt to show announcements of dates past in the past tense, and removing references to dates that have now past as being upcoming.  I hope you will find these proposed changes appropriate.