Remove add-in types we don't collect info for
The Readiness Toolkit doesn't collect VBA, WLL, XLL, XLS or automation add-ins. It used to collect VBA, WLL and XLL in a previous release but this scanning has been removed to remove a dependency on a Windows component. XLS and automation add-ins were never collected. COM (including VSTO) is the only add-in type that is collected.