Update overview-of-shared-computer-activation-for-office-365-proplus.md
Office Deployment Support Engineers have been seeing issues with customers activating Office 365 with single user licensing and then changing the installation to use SCA.  When this is done, the single user license is not automatically removed and the installation will continue to use the single user license.   this is not an issue until the license is up for renewal.   the office installation will see that that SCA is enabled, and then it will see that there is a single user license and it will not renew.   Action customers need to take is to "reset" office back to default installation state by removing the single user license.