Provide details about specificity of C2R package
in the "Deployment differences", we need to specify we provide only one C2R package/source that contains all the Office products (Office subscription, visio and project subscription + VL …) so we do not provide standalone sources for the different product, like we use to provide in MSI. This means the installation package for products, language packs and proofing tools is the C2R package. 
This is needed to understand the size of the source customer will deal with when deploying Office or other Office product.