Update use-the-readiness-toolkit-to-assess-application-compatibility-for-office-365-pro.md
Update the note in the section called "Getting readiness information for multiple users in an enterprise".
It was referring that you need to deploy the tool with the user context to scan MRU. this is wrong. you need user context when you launch the tool to perform a scan of the MRU. not to install it. Most enterprise customer do not provide admin privilege to the users. I provide a way to create a package to generate the scan without installing the tool.