Update office-deployment-tool-configuration-options.md
When SourcePath is pointing to a file path, the value should not include the actually folder that the data have been downloaded. This will result into an error from the setup, which is not useful on systems that do not have access to internet. Systems that do have access to internet will skip the error and download the data from CDN. In a mass installation that may lead into network bandwidth issues.