Fix learn-site-link-broken for site-relative links to images
Image references should always use file-relative links, not site-relative links. A valid site-relative links to an image receive a false negative, and therefore, removal or renaming of the image file won't result in proper reporting. A broken file-relative link is a build warning, but a broken site-relative link to learn.microsoft.com is only a build suggestion.