Update places-analytics.md
Added a section to enable buildings for analytics. This is a new requirement which got done recently and we want to update the documentation to reflect the same. Kindly confirm with me before releasing this article out. I would want to make sure the respective feature is available to customers as we release this.